requirements.txt
langchain
langchain-community
langchain-openai
python-dotenv
pydantic
serpapi
langmem
google-search-results